advanced-agent/pyproject.toml
[project]
name = "advanced-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "firecrawl-py>=4.23.0",
    "langchain>=1.2.15",
    "langchain-ollama>=1.1.0",
    "langgraph>=1.1.9",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
]
